requirements.txt
streamlit
newspaper3k
joblib
nltk
scikit-learn
pandas
